# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IP Management OpenApi Documentation",
        "description": "IP Management application swagger/openapi documentation",
        "contact": {
            "email": "salihanmridha@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "IP Management Server For Common"
        },
        {
            "url": "http://127.0.0.1/api/",
            "description": "IP Management Server For Common IP"
        },
        {
            "url": "http://localhost:8000/api/",
            "description": "IP Management Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login an user",
                "description": "Login an user and Returns json response",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponseDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Email and Password doesn't match"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ip-address": {
            "get": {
                "tags": [
                    "IpAddress"
                ],
                "summary": "Get list of ip addresses",
                "description": "Returns list of ip addresses",
                "operationId": "getIpAddressList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IpAddressResponseDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "IpAddress"
                ],
                "summary": "Store new ip address",
                "description": "Store ip address data",
                "operationId": "store_ip_address",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreIpAddressRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoreIpAddressResponseDTO"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentResponseDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/ip-address/{id}": {
            "get": {
                "tags": [
                    "IpAddress"
                ],
                "summary": "Get ip information",
                "description": "Returns ip data",
                "operationId": "getIpById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ip address id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IpAddressShowResponseDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "404 not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpNotFoundResponseDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "IpAddress"
                ],
                "summary": "Update existing ip",
                "description": "Returns updated ip data",
                "operationId": "updateIp",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "IP id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateIpAddressRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IpAddressShowResponseDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "404 not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpNotFoundResponseDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentResponseDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "HttpForbiddenResponseDTO": {
                "title": "Http forbidden Response DTO",
                "description": "Http forbidden (403 code) Response DTO",
                "properties": {
                    "status": {
                        "title": "status",
                        "type": "integer",
                        "example": 403
                    },
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "title": "error",
                        "type": "string",
                        "example": "You dont have permission to take this action"
                    }
                },
                "type": "object"
            },
            "HttpNotFoundResponseDTO": {
                "title": "Http not found Response DTO",
                "description": "Http not found (404 code) Response DTO",
                "properties": {
                    "status": {
                        "title": "status",
                        "type": "integer",
                        "example": 404
                    },
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "title": "error",
                        "type": "string",
                        "example": "Something not found"
                    }
                },
                "type": "object"
            },
            "IpAddressResponseDTO": {
                "title": "IpAddress Response DTO",
                "description": "IpAddress response DTO",
                "properties": {
                    "status": {
                        "title": "status",
                        "type": "integer",
                        "example": 200
                    },
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "title": "message",
                        "type": "string",
                        "example": "Data successfully retrieve"
                    },
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "properties": {
                                "ip_address": {
                                    "type": "string",
                                    "example": "202.92.249.111,"
                                },
                                "comment": {
                                    "type": "string",
                                    "example": "gifts.ad-group.com.au"
                                },
                                "user_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "created_at": {
                                    "type": "string",
                                    "example": "2023-10-14T03:57:39.000000Z"
                                },
                                "updated_at": {
                                    "type": "string",
                                    "example": "2023-10-14T03:57:39.000000Z"
                                },
                                "user": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Salihan Mridha"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "salihanmridha@gmail.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "IpAddressShowResponseDTO": {
                "title": "Todo show Response DTO",
                "description": "Todo show response DTO",
                "properties": {
                    "status": {
                        "title": "status",
                        "type": "integer",
                        "example": 200
                    },
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "title": "message",
                        "type": "string",
                        "example": "0.0.0.0 ip found"
                    },
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "properties": {
                                "ip_address": {
                                    "type": "string",
                                    "example": "202.92.249.111,"
                                },
                                "comment": {
                                    "type": "string",
                                    "example": "gifts.ad-group.com.au"
                                },
                                "user_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "created_at": {
                                    "type": "string",
                                    "example": "2023-10-14T03:57:39.000000Z"
                                },
                                "updated_at": {
                                    "type": "string",
                                    "example": "2023-10-14T03:57:39.000000Z"
                                },
                                "user": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Salihan Mridha"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "salihanmridha@gmail.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "LoginRequestDTO": {
                "title": "Login Request DTO",
                "description": "Login Request DTO",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "salihanmridha@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password of the user",
                        "type": "string",
                        "example": "12345678"
                    }
                },
                "type": "object"
            },
            "LoginResponseDTO": {
                "title": "Login Response DTO",
                "description": "Login response DTO",
                "properties": {
                    "status": {
                        "title": "status",
                        "type": "integer",
                        "example": 200
                    },
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "title": "message",
                        "type": "string",
                        "example": "Login successfully!"
                    },
                    "data": {
                        "title": "data",
                        "properties": {
                            "access_token": {
                                "type": "string",
                                "example": "your-access-token"
                            },
                            "user": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Salihan Mridha"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "salihanmridha@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StoreIpAddressRequestDTO": {
                "title": "Store Ip address Request DTO",
                "description": "Store Ip address Request DTO",
                "required": [
                    "ip_address",
                    "comment"
                ],
                "properties": {
                    "ip_address": {
                        "title": "ip_address",
                        "description": "Valid IP address",
                        "type": "string",
                        "example": "121.28.229.28"
                    },
                    "comment": {
                        "title": "comment",
                        "description": "Comment of IP address",
                        "type": "string",
                        "example": "Comment for ip 121.28.229.28"
                    }
                },
                "type": "object"
            },
            "StoreIpAddressResponseDTO": {
                "title": "Store IpAddress Response DTO",
                "description": "Store IpAddress response DTO",
                "properties": {
                    "status": {
                        "title": "status",
                        "type": "integer",
                        "example": 201
                    },
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "title": "message",
                        "type": "string",
                        "example": "127.0.0.1 ip successfully created."
                    },
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "properties": {
                                "ip_address": {
                                    "type": "string",
                                    "example": "202.92.249.111,"
                                },
                                "comment": {
                                    "type": "string",
                                    "example": "gifts.ad-group.com.au"
                                },
                                "user_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "created_at": {
                                    "type": "string",
                                    "example": "2023-10-14T03:57:39.000000Z"
                                },
                                "updated_at": {
                                    "type": "string",
                                    "example": "2023-10-14T03:57:39.000000Z"
                                },
                                "user": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Salihan Mridha"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "salihanmridha@gmail.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "UnprocessableContentResponseDTO": {
                "title": "Unprocessable Content Response DTO",
                "description": "Unprocessable Content (422 code) Response DTO",
                "properties": {
                    "status": {
                        "title": "status",
                        "type": "integer",
                        "example": 422
                    },
                    "success": {
                        "title": "success",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "title": "message",
                        "type": "string",
                        "example": "Requested data validation failed."
                    },
                    "errors": {
                        "title": "errors",
                        "properties": {
                            "key": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Validation error shows for the key"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UpdateIpAddressRequestDTO": {
                "title": "Update Ip address Request DTO",
                "description": "Update Ip address Request DTO",
                "required": [
                    "comment"
                ],
                "properties": {
                    "comment": {
                        "title": "comment",
                        "description": "Comment of IP address",
                        "type": "string",
                        "example": "Comment for ip 121.28.229.28"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}